http2: make usable
HTTP/2 support is now able to serve the test server, complete with
websockets, from a single vhost.

 - This works the same with both OpenSSL and mbedTLS.

 - POST is now wired up and works.

 - CGI is wired up and works.

 - Redirect is adapted and works

 - lwsws works.

 - URI urldecode, sanitation and argument parsing wired up for :path

valgrind clean (aside from openssl-style false uninit data usage in mbedtls send occasionally)

h2spec reports:

$ h2spec  -h 127.0.0.1 -p 7681 -t -k -o 1
...
145 tests, 144 passed, 1 skipped, 0 failed"

Incorporates:

 - "https://github.com/warmcat/libwebsockets/pull/1039
	Fixes issue with -Werror=unused-variable flag

 - https://github.com/sartura/libwebsockets/commit/2c843a1395bdca5c23b40d989f508fc96aa5ba3b
	ssl: fix infinite loop on client cert verification failure


Caused and fixes Coverity 184887 - 184892

+++ plugins/ssh-base/crypto/ed25519.c
@@ -194,7 +194,7 @@ int crypto_sign_ed25519_open(
 	  return -1;
   }
 
-  get_hram(hram,sm,pk,m,smlen);
+  get_hram(hram,sm,pk,m, (size_t)smlen);
 
   sc25519_from64bytes(&schram, hram);
